# proto/calc.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "calc.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Calc"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/increment": {
      "post": {
        "operationId": "Calc_Increment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoNumResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoNumRequest"
            }
          }
        ],
        "tags": [
          "Calc"
        ]
      }
    },
    "/v1/work": {
      "post": {
        "operationId": "Calc_Work",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoWorkResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoWorkRequest"
            }
          }
        ],
        "tags": [
          "Calc"
        ]
      }
    }
  },
  "definitions": {
    "protoNumRequest": {
      "type": "object",
      "properties": {
        "val": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protoNumResponse": {
      "type": "object",
      "properties": {
        "val": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protoWorkRequest": {
      "type": "object"
    },
    "protoWorkResponse": {
      "type": "object",
      "properties": {
        "reply": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}